ml_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
opencv-python==4.8.1.78
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
